+++ src/main/java/com/marginallyclever/robotOverlord/mantisRobot/tool/MantisTool.java
@@ -47,5 +47,5 @@ public abstract class MantisTool extends PhysicalObject {
 	
 	public void updateGUI() {}
 	
-	public void update(float dt) {}
+	public void update(double dt) {}
 }
